Migrate other 2023 challenges
