+++ aocd_python/2023/day1.py
