+++ aocd_python/2023/day6.py
@@ -1,0 +1,6 @@
+def part1(data):
+    pass
+
+
+def part2(data):
+    pass
